fixes to shell scripts

+++ deploy.sh
@@ -4,6 +4,7 @@ files=(
     "vim/init.vim, ~/.config/nvim" 
     ".zshenv, ~"
     ".zshrc, ~"
+	".tmux.conf, ~"
 )
 
 for paths in "${files[@]}"; do
@@ -11,6 +12,6 @@ for paths in "${files[@]}"; do
     disk_path=$(echo ${paths} | cut -d "," -f 2 | tr -d '[:space:]')/$(basename $repo_path)
     disk_path=${disk_path/#\~/$HOME} #replace tilde with current home folder
 
-	mv $disk_path "${disk_path}.backup"
+	[ -f $disk_path ] && mv $disk_path "${disk_path}.backup"
 	cp $repo_path $disk_path
 done
